% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{basicVisualization}
\alias{basicVisualization}
\title{Get Basic Visualizations for Simulations}
\usage{
basicVisualization(data)
}
\arguments{
\item{data}{Simulation data -- the output of \code{\link{generateData}}.}
}
\value{
A series of plots that can be indexed into via the \code{$} operator.
  Specifically, the following plots can be accessed: \itemize{ \item
  \emph{$Prop1Plot:} A bar plot with the proportions of actions (right or
  left) in State FC. \item \emph{$Prop2Plot:} A bar plot with the proportions
  of actions (right or left) in States A and B (faceted appropriately). \item
  \emph{$State1ActPlot:} A line plot of actions (right or left) in State FC
  over time. \item \emph{$State2ActPlot:} A line plot of actions (right or
  left) in States A and B (faceted appropriately) over time. \item
  \emph{$ImagePropQPlot:} A line plot showing the probability of receiving a
  reward and the average Q value associated with each image over time. \item
  \emph{$Qtable:} A data frame containing the final Q table for this
  simulation. \item \emph{$QRorL:} A data frame that has the Q values
  associated with going right or left from the first state. }
}
\description{
Get Basic Visualizations for Simulations
}
