% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice.R
\name{choice}
\alias{choice}
\title{Choose an Action}
\usage{
choice(state, epsilon = 0.1, tau = 0.08, softmax = TRUE)
}
\arguments{
\item{state}{The state the agent is currently in.}

\item{epsilon}{The epsilon to be used in epsilon-greedy policy choices.}

\item{tau}{The tau (temperature) to be used in softmax policy choices.}

\item{softmax}{Logical: TRUE if softmax policy decisions should be used;
FALSE if epsilon-greedy policy decisions should be used. By default,
softmax is used.}
}
\value{
A string depicting an action (left or right) to take in the current
  state.
}
\description{
Choose an Action
}
