% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate Task Simulations}
\usage{
generateData(
  numIter = 201,
  x = 10,
  alpha = 0.1,
  gamma = 0.9,
  epsilon = 0.1,
  tau = 0.08,
  softmax = TRUE
)
}
\arguments{
\item{numIter}{The number of trials per session.}

\item{x}{The amount of simulations to be done. This is the extent to which
model-based activity occurs.}

\item{alpha}{The learning rate alpha.}

\item{gamma}{The temporal discounting factor gamma.}

\item{epsilon}{The epsilon to be used in epsilon-greedy policy choices.}

\item{tau}{The tau (temperature) to be used in softmax policy choices.}

\item{softmax}{Logical: TRUE if softmax policy decisions should be used;
FALSE if epsilon-greedy policy decisions should be used. By default,
softmax is used.}
}
\value{
If \code{x} is greater than 0, so there are some model-based
  simulations, then this function returns a list with two elements containing
  real (element 1) and simulated (element 2) data. These can be passed into
  the \code{\link{processSimData}} function to concatenate into one
  dataframe. If \code{x} is equal to 0, so there are no model-based
  simulations, then this function returns a dataframe with real experience
  data.
}
\description{
Running this function is equivalent to running one agent on the two-stage
Markov decision task.
}
