% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticRegression.R
\name{getLogPreds}
\alias{getLogPreds}
\title{Get the Logistic Regression Predictions}
\usage{
getLogPreds(data)
}
\arguments{
\item{data}{The data can \emph{either} be the output of
\code{\link{logSetup}} \emph{or} the output of \code{\link{generateData}}.
If the latter, make sure that manipulateData is called on pure Q-learning
simulations. Similarly, for DynaQ simulations, make sure the data has been
processed and the Q table has been removed.}
}
\value{
The original (modified with \code{\link{logSetup}}) data frame plus
  two columns \code{predNoStay} and \code{predStay}.
}
\description{
This function takes in simulation data (see parameter data for more
information) and calls the \code{\link{logSetup}} function if necessary
before calling the \code{\link{getLogFit}} function to fit a logistic
regression model predicting the probability of staying based on whether the
last trial was rewarded or not, whether the last trial had a common or rare
transition, and the interaction between the two. It then uses the predict
function and binds the predictions to the original (manipulated with
\code{\link{logSetup}}) dataframe.
}
