% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewards.R
\name{getReward}
\alias{getReward}
\title{Check for Reward Outcome}
\usage{
getReward(state)
}
\arguments{
\item{state}{The state, or image, for which a reward might be received.}
}
\value{
Binary reward (1 or 0) based off of the relevant state probability.
}
\description{
the getReward function determines whether a reward is delivered (1) or not
(0) based off of a binary payoff with probability either 0.2 or 0.8 that
have, on each trial, a little bit of gaussian noise following the normal
distribution with mean 0 and SD 0.025. See \code{\link{updateRewardProb}}
}
