% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{getSecondTransition}
\alias{getSecondTransition}
\title{Get Transition from Second to Third State}
\usage{
getSecondTransition(state, action)
}
\arguments{
\item{state}{Current second state.}

\item{action}{Action taken in second state.}
}
\value{
Final state and reward.
}
\description{
\code{getSecondTransition} takes in the current second state and action
chosen in that state and determines the final state and reward. This function
does not change whether or not the experience is real or simulated.
}
