% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticRegression.R
\name{logSetup}
\alias{logSetup}
\title{Manipulate the Data for Logistic Regression}
\usage{
logSetup(data)
}
\arguments{
\item{data}{The output of \code{\link{generateData}}; if pure Q-learning, it manipulateData
should be called first. If it's DynaQ, the data should be processed and the
Q table removed.}
}
\value{
A dataframe with 23 columns. New ones include \code{lastAction1}
  (numeric) that tracks the action taken in the previous trial,
  \code{lastReward} that tracks whether the last trial was rewarded,
  (numeric), and \code{stay} (factor) which tracks whether trial n and n+1
  have the same first action.
}
\description{
This function takes in the output of \code{\link{generateData}} (see parameter data for
specifics) and sets it up for logistic regression by mutating the dataframe
so reward and transition refer to the last trial's reward and transition type
and the stay column means that the (first) action on trial n is equal to the
(first) action on trial n = 1. This function gets called in
\code{\link{getLogFit}} and \code{\link{getLogPreds}} as needed.
}
