% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSimulations.R
\name{manipulateData}
\alias{manipulateData}
\title{Manipulate Data}
\usage{
manipulateData(data)
}
\arguments{
\item{data}{The data to be manipulated. Typically, the output of
\code{\link{generateData}}, though it could be the output of
\code{\link{oneTrial}} or \code{\link{simModel}.}}
}
\value{
Manipulated data with 21 columns and however many trials were run
  (just real or with real and simulated experience combined).
}
\description{
This function manipulates the output of \code{\link{generateData}}. It
recodes actions as binary (right == 1, left == 0), calculates the proportion
of actions right or left as of each trial, determines if a transition is
common or not (inherent in the task structure), and filters the data to
remove the Qtable while retaining meta data such as the temporal discounting
factor (gamma) learning rate for simulated data (alpha), and the probability
of receiving a reward for each image.
}
