% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFree.R
\name{oneTrial}
\alias{oneTrial}
\title{Perform One Model-Free Trial}
\usage{
oneTrial(alpha = 0.1, gam = 0.9, epsilon = 0.1, tau = 0.08, softmax = TRUE)
}
\arguments{
\item{alpha}{The learning rate alpha.}

\item{gam}{The temporal discounting factor gamma.}

\item{epsilon}{The epsilon to be used in epsilon-greedy policy choices.}

\item{tau}{The tau (temperature) to be used in softmax policy choices.}

\item{softmax}{Logical: TRUE if softmax policy decisions should be used;
FALSE if epsilon-greedy policy decisions should be used. By default,
softmax is used.}
}
\value{
A tibble with 8 rows and 18 columns. The 8 rows contain identical
  information \strong{except for the Qtable column.} They contain information
  about the states, actions, and rewards for one trial as well as meta data
  including the temporal discounting factor (gamma) learning rate (alpha),
  choice policy parameters (epsilon and tau), and probability of receiving a
  reward for each image.
}
\description{
The \code{oneTrial} function simulates one trial of the two-stage Markov task using model-free Q-learning.
}
