% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSimulations.R
\name{processSimData}
\alias{processSimData}
\title{Process Simulated Data}
\usage{
processSimData(data, x, removeTable)
}
\arguments{
\item{data}{The output of \code{\link{generateData}} if it's not pure
model-free.}

\item{x}{The number of simulations per real trial.}

\item{removeTable}{Logical: TRUE if the Qtable should be removed. FALSE if it
should not be removed.}
}
\value{
A dataframe of real and simulated data in the appropriate order.

If removeTable = TRUE, the output of this does not contain the Q table and
can be passed into functions such as stayProbabilityPlot.

If removeTable = FALSE, the output of this contains the Q table and can be passed into functions such as getPlots.
}
\description{
This function takes in the output of generateData if x is greater than 0 and
returns a sorted version that has real and simulated experiences in the
appropriate order, either with or without the Qtable.
}
