% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelBased.R
\name{randomPrevious}
\alias{randomPrevious}
\title{Get a Random Value from an Existing Dataframe}
\usage{
randomPrevious(df, name)
}
\arguments{
\item{df}{A dataframe to get the random value from}

\item{name}{The column name of the dataframe}
}
\value{
A random value from \code{df$name} as a character string.
}
\description{
Returns a character string of a random value from a supplied dataframe and
column name.
}
