% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewards.R
\name{randRewardProb}
\alias{randRewardProb}
\title{Random Reward Probability}
\usage{
randRewardProb()
}
\value{
a random number between 0.2 and 0.8
}
\description{
randRewardProb function generates a random number based on a uniform
distribution between 0.2 and 0.8.
}
\examples{
\dontrun{
x <- dynqrandRewardProb() #x will be between 0.2 and 0.8
}

}
