% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSimulations.R
\name{removeTable}
\alias{removeTable}
\title{Remove Qtable From Data}
\usage{
removeTable(data)
}
\arguments{
\item{data}{Any dataframe that has a Qtable column, such as the output of
\code{\link{oneTrial}} or \code{\link{simModel}}.}
}
\value{
The unique values of the data (minus the Qtable).
}
\description{
This function filters the inputted dataframe so we only get every 8th row.
Because there are 8 rows in the Q table, which we have in the output of
\code{\link{generateData}}, everything trial's information is repeated 8
times. This function uses a dplyr::filter for \code{(row_number() + 1 %% 8 ==
1}, which returns the 1st, 9th, 17th, 25th, etc. rows which happen to be the
first one for the new action set.
}
