% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.R
\name{setupQtable}
\alias{setupQtable}
\title{Initialize Q-value Information}
\usage{
setupQtable()
}
\value{
\code{Q$Table}, \code{Q$R}, \code{Q$L}, and two vectors used in the
\code{\link{updateQsecondState}} and \code{\link{updateQthirdState}} functions.
}
\description{
setupQtable initializes the \code{Q$Table}, including the \code{Q$R} and
\code{Q$L} values, plus the comparison vectors used to update Q-values in the
\code{\link{updateQsecondState}} and \code{\link{updateQthirdState}} functions.
}
