% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewards.R
\name{setupRewards}
\alias{setupRewards}
\title{Setup Rewards}
\usage{
setupRewards()
}
\value{
Returns the initial reward probabilities for each image
}
\description{
setupRewards function initializes the reward probabilities for each image at
the beginning of a session.
}
\details{
The four probabilities are initialized as global variables:
\itemize{
    \item probAR
    \item probAL
    \item probBR
    \item probBL
}
}
