% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{setupTransFunction}
\alias{setupTransFunction}
\title{Setup Transition Model}
\usage{
setupTransFunction()
}
\value{
Returns the initialized simulation counter \code{trans$numSims}, the
  transition function \code{trans$Function}, the transition table
  \code{trans$Table} which keeps track of the
  State/Action/StatePrime/Reward/source of experience (simulated or not), and
  the more detailed \code{trans$Track}, which keeps the running proportion of
  states observed as a result of the first action/transition.
}
\description{
Setup Transition Model
}
