% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelBased.R
\name{simModel}
\alias{simModel}
\title{Perform One Model-Based Trial}
\usage{
simModel(trialData, modelAlpha = 0.1, gam = 0.9, epsilon = 0.1, tau = 0.08, x)
}
\arguments{
\item{trialData}{The output of \code{\link{oneTrial}} (i.e. the last real
experience)}

\item{modelAlpha}{The learning rate from simulated data.}

\item{gam}{The temporal discounting factor, gamma.}

\item{epsilon}{The epsilon to be used in epsilon-greedy policy choices.}

\item{tau}{The tau (temperature) to be used in softmax policy choices.}

\item{x}{The amount of simulations to be done. This is used to track the
total number performed via the \code{\link{updateTransFunction}}.}
}
\value{
A tibble with 8 rows and 16 columns. The 8 rows contain identical
  information \strong{except for the Qtable column.} They contain information
  about the states, actions, and rewards for one trial as well as meta data
  including the temporal discounting factor (gamma) learning rate for
  simulated data (alpha), and the probability of receiving a reward for each
  image.
}
\description{
The \code{simModel} function simulates one trial of the two-stage Markov task
using a model-based approach (with a transition model), whose action values
get updated according to the Q-learning. The model-based simulations are
based on random previously visited states and previously taken actions. The x
parameter is the number of simulations that are run, in line with the Dyna
architecture.
}
