% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{simTransitionPlot}
\alias{simTransitionPlot}
\title{Plot the Transition Model's Estimates Over Time}
\usage{
simTransitionPlot(trackedTransData, text = TRUE, hline = TRUE)
}
\arguments{
\item{trackedTransData}{The \code{trans$Track} for a given simulation.}

\item{text}{Logical: TRUE will display text on the plot describing the true
transition probabilities. FALSE will disable this. Default is TRUE.}

\item{hline}{Logical: TRUE will display a horizontal line on the plot at the
true transition probabilities. FALSE will disable this. Default is TRUE.}
}
\value{
A plot showing the model's estimates of transition probabilities over time.
}
\description{
Plot the Transition Model's Estimates Over Time
}
