% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choice.R
\name{smaxAction}
\alias{smaxAction}
\title{Action according to Softmax}
\usage{
smaxAction(smax)
}
\arguments{
\item{smax}{The softmax argument (specifically the softmax probability of going right -- \code{smaxRight}).}
}
\value{
An action based on a random binomial distribution with probability smax.
}
\description{
Action according to Softmax
}
