% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{stayDistributionPlot}
\alias{stayDistributionPlot}
\title{Plot the Distribution of Stay Probabilities for Multiple Agents}
\usage{
stayDistributionPlot(data)
}
\arguments{
\item{data}{Data is a single dataframe that has had the
\code{\link{manipulateData}} function applied to it. Should be a dataframe
with multiple agents.}
}
\value{
A histogram showing the distribution of stay probabilities amongst
  multiple agents.
}
\description{
This function plots a histogram of the distribution of stay probabilities for
each agent. The plot is faceted by transition type (common or rare) and
reward (yes or no).
}
