% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{summaryTable}
\alias{summaryTable}
\title{Create a Summary Table of the Stay Probabilities}
\usage{
summaryTable(data)
}
\arguments{
\item{data}{Data is a single dataframe that has had the
  \code{\link{manipulateData}} function applied to it. Should be a dataframe
  with multiple agents.

  alpha gamma reward transition meanPredStay stdPredStay minPredStay
  maxPredStay}
}
\value{
A dataframe with 8 coiumns and variable rows depending on the
  combinations of the learning rate, alpha, and temporal discount factor,
  gamma:

  \itemize{ \item \emph{alpha:} Numeric -- learning rate, alpha. \item
  \emph{gamma:} Numeric -- temporal discount factor, gamma. \item
  \emph{reward:} Factor -- yes or no. \item \emph{transition:} Factor -- common
  or rare transition type. \item \emph{meanPredStay:} Numeric -- the mean
  predicted probability of staying. \item \emph{stdPredStay:} Numeric -- the
  standard deviation of the predicted probability of staying. \item
  \emph{minPredStay:} Numeric -- the minimum predicted probability of staying.
  \item \emph{maxPredStay:} Numeric -- the maximum predicted probability of
  staying. }
}
\description{
This function creates a summary table of the mean, standard deviation,
minimum, and maximum stay probabilities based on the predictions from a
logistic regression. It is specific to each reward and transition structure
combination.
}
