% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.R
\name{updateQRL}
\alias{updateQRL}
\title{Update QR and QL}
\usage{
updateQRL(state, action, alpha, gamma)
}
\arguments{
\item{state}{The state the agent is currently in.}

\item{action}{The action (right or left) an agent performs.}

\item{alpha}{The learning rate alpha.}

\item{gamma}{The temporal discounting factor gamma.}
}
\value{
Updated \code{Q$R} or \code{Q$L} values.
}
\description{
The \code{updateQRL} function will take in the action an agent performs and
update either \code{Q$R} or \code{Q$L} as appropriate.
}
