% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.R
\name{updateQsecondState}
\alias{updateQsecondState}
\title{Update Q-values at Second State}
\usage{
updateQsecondState(state, action, alpha, gamma)
}
\arguments{
\item{state}{The state the agent is currently in.}

\item{action}{The action (right or left) an agent performs.}

\item{alpha}{The learning rate alpha.}

\item{gamma}{The temporal discounting factor gamma.}
}
\value{
Updated second state Q-value
}
\description{
This function takes the \code{Q$stateActionPair} dataframe initialized in the
background and filters it to pull the Q-table index that matches the
state-action pair. It then updates the Q-value according to the Q-learning
algorithm.
}
