% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.R
\name{updateQtable}
\alias{updateQtable}
\title{Update Q Table}
\usage{
updateQtable(state, action, reward, alpha = 0.1, gamma = 0.9)
}
\arguments{
\item{state}{The state the agent is currently in.}

\item{action}{The action (right or left) an agent performs.}

\item{reward}{The reward an agent receives at the current state.}

\item{alpha}{The learning rate alpha.}

\item{gamma}{The temporal discounting factor gamma.}
}
\value{
Updated Q-values.
}
\description{
This function updates the Q-table values by calling the functions
\code{\link{updateQRL}} which updates the \code{Q$R} and \code{Q$L} values,
\code{\link{updateQsecondState}}, which updates the second state Q-values,
and \code{\link{updateQthirdState}}, which updates the third state Q-values.
}
