% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q.R
\name{updateQthirdState}
\alias{updateQthirdState}
\title{Update Q-values at Third State}
\usage{
updateQthirdState(state, reward, alpha, gamma)
}
\arguments{
\item{state}{The state the agent is currently in.}

\item{reward}{The reward an agent receives at the current state.}

\item{alpha}{The learning rate alpha.}

\item{gamma}{The temporal discounting factor gamma.}
}
\value{
Updated third state Q-value
}
\description{
This function takes the \code{Q$thirdStateVec} dataframe initialized in the
background and filters it to pull the Q-table index that matches the
terminal state. It then updates the Q-value according to the Q-learning
algorithm.
}
