% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewards.R
\name{updateRewardProb}
\alias{updateRewardProb}
\title{Update Reward Probabilities}
\usage{
updateRewardProb(imageProb)
}
\arguments{
\item{imageProb}{The current probability of receiving a reward for a given
image.}
}
\value{
Updated probability
}
\description{
Update reward probabilities according to Solway et al. (2019) methods. It
takes in an image's reward probability, and returns a final probability to be
reassigned to the inputted parameter according to the above description.where
probabilities shift according to random gaussian noise with mean 0 and
standard deviation of 0.025. Reflecting boundaries are at 0.8 and 0.2.
}
\details{
According to personal communication with Solway: Each of the four
  second-stage images had independent probabilities.

  \itemize{ \item A random probability was chosen for each for trial 1 in the
  valid range. \item The probability for trial n+1 was: prob trial n + N(0,
  sd) \item There were reflecting boundaries, meaning, for example, if prob
  on trial n was 0.7 and you drew +0.3, you would end up at 0.6 for trial n+1
  (go up 0.1 to the 0.8 upper boundary, then down 0.2).

  }
}
