% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitions.R
\name{updateTransFunction}
\alias{updateTransFunction}
\title{Update Transition Model}
\usage{
updateTransFunction(action, statePrime, alpha, real = NULL, x)
}
\arguments{
\item{action}{The action taken in the initial state}

\item{statePrime}{The state reached as a result of taking action a in the
initial state.}

\item{alpha}{The learning rate for simulated data.}

\item{real}{The previously experienced "real" S, A, R, S' combination.}

\item{x}{The number of simulations to perform.}
}
\value{
Updates the trans environment objects with the latest values.
}
\description{
The updateTransFunction updates the transition model for simulated
experience. It updates the following: \itemize{ \item \code{trans$numSims}
which keeps track of the number of simulations currently run. \item
\code{trans$Function} which contains the model's estimate of transitioning to
different states from FC. \item The transition table \code{trans$Table}
which keeps track of the State/Action/StatePrime/Reward/source of experience
(simulated or not). \item \code{trans$Track} which keeps the
running proportion of states observed as a result of the first
action/transition (more detailed \code{trans$Table}). }
}
